requirements.txt
# Waymo Waymax Simulator
waymo-waymax @ git+https://github.com/waymo-research/waymax.git@main
# JAX with CUDA 12 support
jax[cuda12]
# Neural Network Libraries
flax
distrax
einops
# Utilities
hydra-core
matplotlib<3.10
tensorboardX
psutil
pandas
seaborn
